ISSUE #2763 - move modelList func to commons
